Set up a selection of translations for test
